Move docker-compose.yml validation to utils
